requirements.txt

chromadb
streamlit
langchain-community
huggingface_hub
sentence-transformers
transformers
